Use Int128 when interpolating
Technically not needed unless the shift is modified to be greater than 16, but this does make it easy to just change the shift in the future.

+++ Exanite.Core/Numerics/Fixed.Log.cs
@@ -40,7 +40,7 @@ public partial struct Fixed
             // Get values and interpolate
             var y0 = Log2Lut[index];
             var y1 = Log2Lut[index + 1];
-            normalizedResult = y0 + (((y1 - y0) * fraction) >> Shift);
+            normalizedResult = (long)(y0 + (((Int128)(y1 - y0) * fraction) >> Shift));
         }
 
         return new Fixed(normalizedResult) - shiftNormalize;
